pyproject.toml
[project]
name = "code_conquest"
version = "0.1.0"
description = "Code Conquest - The Adventure Game"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "arcade>=3.3.2",
    "python-dotenv==1.1.1",
    "supabase==2.17.0",
]

[tool.yapf]
indent_width = 2

[tool.basedpyright]
typeCheckingMode = "standard"
reportIncompatibleMethodOverride = "none"

reportMissingTypeStubs = "none"
reportUnusedCallResult = "none"
reportUnknownArgumentType = "none"
reportUnknownMemberType = "none"
reportPrivateImportUsage = "none"
